Update to v1.5.0
Fairly large code update to make proper package. Lots of linting and
updates to newer Python practices. Added proper tests. Added ability to
get and plot MD areas (cf: #2). Updated tutorial.
