+++ .github/workflows/lint.yaml
@@ -1,0 +1,26 @@
+name: Linting
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  flake8:
+    name: Flake8
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Set up Python 3
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.x
+          cache: 'pip'
+          cache-dependency-path: 'ci/linting_requirements.txt'
+
+      - name: Install linting tools
+        run: python -m pip install -r ci/linting_requirements.txt
+
+      - name: Run flake8
+        run: flake8
