+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2020, Nathan Wendt
+Copyright (c) 2023, Nathan Wendt
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
