+++ ci/linting_requirements.txt
@@ -1,0 +1,22 @@
+flake8
+pycodestyle
+pyflakes
+
+flake8-bugbear
+flake8-builtins
+flake8-comprehensions
+flake8-continuation
+flake8-copyright
+flake8-isort
+isort
+flake8-mutable
+flake8-pie
+flake8-print
+flake8-quotes
+flake8-requirements
+flake8-simplify
+
+flake8-docstrings
+pydocstyle
+
+codespell
