+++ setup.cfg
@@ -1,0 +1,67 @@
+[metadata]
+name = SPCartopy
+version = attr: spcartopy.__version__
+author = Nathan Wendt
+author_email = nathan.wendt@noaa.gov
+description = Add SPC products to cartopy maps.
+license = BSD-3-Clause
+license_file = LICENSE
+
+[build-system]
+requires = ["setuptools"]
+
+[options]
+package_dir=
+    =src
+packages=find:
+setup_requires =
+    setuptools
+install_requires =
+    cartopy
+    fiona
+    matplotlib
+    shapely >= 1.8
+
+[options.packages.find]
+where = src
+
+[options.extras_require]
+lint = 
+    flake8
+    pycodestyle
+    pyflakes
+    flake8-bugbear
+    flake8-builtins
+    flake8-comprehensions
+    flake8-continuation
+    flake8-copyright
+    flake8-isort
+    isort
+    flake8-mutable
+    flake8-pie
+    flake8-print
+    flake8-quotes
+    flake8-requirements
+    flake8-simplify
+    flake8-docstrings
+    pep8-naming
+    pydocstyle
+
+test =
+    pytest
+    pytest-mpl
+
+[flake8]
+max-line-length = 95
+application-import-names = mdgpu
+import-order-style = google
+copyright-check = True
+copyright-author = Nathan Wendt
+inline-quotes = single
+multiline-quotes = double
+docstring-convention = numpy
+select = A B C D E F G H I J K L M N O P Q R S T U V W X Y Z B902
+ignore = F405 W503 RST902 SIM106 PIE786
+per-file-ignores =
+    __init__.py:F401
+    patheffects.py:N803
