--- spcartopy/__init__.py
@@ -1,6 +1,0 @@
-from cartopy import config
-
-import spcartopy.colors
-import spcartopy.feature
-import spcartopy.hatch
-import spcartopy.legends
