+++ src/spcartopy/__init__.py
@@ -1,0 +1,13 @@
+# Copyright (c) 2023 Nathan Wendt.
+# Distributed under the terms of the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+"""SPCartopy: Add SPC products to cartopy maps."""
+
+from cartopy import config
+
+import spcartopy.colors
+import spcartopy.feature
+import spcartopy.hatch
+import spcartopy.legends
+
+__version__ = '1.5.0'
