+++ src/spcartopy/io/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2023 Nathan Wendt.
+# Distributed under the terms of the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+"""SPCartopy I/O module."""
